Write mesher part of texcoord packing. Buggy

+++ terra-meshgen/src/main/java/com/ritualsoftheold/terra/mesher/VoxelMesher.java
@@ -31,6 +31,6 @@ public interface VoxelMesher {
     
     IntList getIndices();
     
-    FloatList getTextureCoords();
+    IntList getTextureCoords();
 
 }
